#126:  All prices and exchange rates should be decimals. Added a ShopInventoryViMo test.

+++ dev/Shop/CurrencyM.cs
@@ -19,5 +19,5 @@ public class CurrencyM
 
     public string Name { get; }
 
-    public double Rate { get; }
+    public decimal Rate { get; }
 }
